added gameApi, added signalR, refactoring
