+++ frontend/package.json
@@ -33,6 +33,7 @@
     ]
   },
   "devDependencies": {
+    "@microsoft/signalr": "^8.0.7",
     "axios": "^1.7.9",
     "jwt-decode": "^4.0.0",
     "react-router-dom": "^7.1.5"
