+++ frontend/src/pages/Home/styles/HomePage.css
@@ -1,0 +1,66 @@
+.room-list-container {
+    padding: 20px;
+    text-align: center;
+    display: flex;
+    flex-direction: column;
+}
+
+.room-list-container h2 {
+    font-size: 2rem;
+    margin-bottom: 20px;
+}
+
+.room-list-container .rooms {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+    gap: 20px;
+    justify-items: center;
+}
+
+.room-card {
+    background-color: #f9f9f9;
+    border: 1px solid #ddd;
+    border-radius: 8px;
+    padding: 15px;
+    width: 100%;
+    max-width: 350px;
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+    transition: transform 0.3s ease;
+}
+
+.room-card:hover {
+    transform: scale(1.05);
+}
+
+.room-card h3 {
+    font-size: 1.5rem;
+    margin-bottom: 10px;
+}
+
+.room-card p {
+    font-size: 1rem;
+    color: #555;
+    margin: 5px 0;
+}
+
+.room-card button {
+    background-color: #4CAF50;
+    color: white;
+    border: none;
+    padding: 10px;
+    border-radius: 5px;
+    cursor: pointer;
+    font-size: 1rem;
+    width: 100%;
+    margin-top: 15px;
+}
+
+.room-card button:hover {
+    background-color: #45a049;
+}
+
+.loading-text {
+    font-size: 1.2rem;
+    color: #888;
+    margin-top: 20px;
+}
